pyhub/DAY-02 (2025-06-18)/실습/requirements.txt
# 라이브러리 의존성 명시
#  - 의존성 (dependencies)
#    이 애플리케이션 구동에 꼭 필요하다.
django-environ

pyhub-llm[all]
